+++ chipsec/library/uefi/search.py
@@ -119,7 +119,7 @@ def check_rules(efi: EFI_SECTION, rules: Dict[str, Any], entry_name: str, _log: 
             match_mask |= MATCH_NAME
         if ('guid' in rule) and (rule['guid'] != ''):
             match_mask |= MATCH_GUID
-            if type(rule['guid']) == str:
+            if isinstance(rule['guid'], str):
                 rule['guid'] = UUID(rule['guid'])
         if ('regexp' in rule) and (rule['regexp'] != ''):
             match_mask |= MATCH_REGEXP
