+++ chipsec/utilcmd/uefi_cmd.py
@@ -64,7 +64,7 @@ from chipsec.library.file import write_file, read_file
 from chipsec.library.uefi.spi import decode_uefi_region, modify_uefi_region, compress_image, CMD_UEFI_FILE_REPLACE
 from chipsec.library.uefi.spi import CMD_UEFI_FILE_INSERT_AFTER, CMD_UEFI_FILE_INSERT_BEFORE, CMD_UEFI_FILE_REMOVE
 from chipsec.hal.common.uefi import UEFI
-from chipsec.library.uefi import SECURE_BOOT_KEY_VARIABLES, get_attr_string
+from chipsec.library.uefi.variables import SECURE_BOOT_KEY_VARIABLES, get_attr_string
 from chipsec.library.uefi.sleep_states import parse_script
 from chipsec.library.uefi.varstore import parse_efivar_file, decode_EFI_variables, identify_EFI_NVRAM, parse_EFI_variables
 from chipsec.library.uefi.fv import get_guid_bin, assemble_uefi_file, assemble_uefi_section, assemble_uefi_raw
